+++ rust/main/hyperlane-base/src/settings/chains.rs
@@ -20,10 +20,8 @@ use hyperlane_metric::prometheus_metric::ChainInfo;
 use hyperlane_operation_verifier::ApplicationOperationVerifier;
 
 use hyperlane_cosmos::{
-    self as h_cosmos, delivery_indexer, dispatch_indexer, rpc::CosmosWasmRpcProvider,
-    CosmosProvider, Signer,
+    self as h_cosmos, cw::CwQueryClient, native::ModuleQueryClient, CosmosProvider,
 };
-use hyperlane_cosmos_native::{self as h_cosmos_native, CosmosNativeProvider};
 use hyperlane_ethereum::{
     self as h_eth, BuildableWithProvider, EthereumInterchainGasPaymasterAbi, EthereumMailboxAbi,
     EthereumReorgPeriod, EthereumValidatorAnnounceAbi,
@@ -176,7 +174,7 @@ pub enum ChainConnectionConf {
     /// Starknet configuration.
     Starknet(h_starknet::ConnectionConf),
     /// Cosmos native configuration
-    CosmosNative(h_cosmos_native::ConnectionConf),
+    CosmosNative(h_cosmos::ConnectionConf),
     /// Radix configuration
     Radix(h_radix::ConnectionConf),
 }
@@ -307,7 +305,7 @@ impl ChainConf {
                 Ok(Box::new(provider) as Box<dyn HyperlaneProvider>)
             }
             ChainConnectionConf::Cosmos(conf) => {
-                let provider = build_cosmos_provider(self, conf, metrics, &locator, None)?;
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, None)?;
                 Ok(Box::new(provider) as Box<dyn HyperlaneProvider>)
             }
             ChainConnectionConf::Starknet(conf) => {
@@ -363,9 +361,9 @@ impl ChainConf {
             }
             ChainConnectionConf::Cosmos(conf) => {
                 let signer = self.cosmos_signer().await.context(ctx)?;
-                let provider = build_cosmos_provider(self, conf, metrics, &locator, signer)?;
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, signer)?;
 
-                h_cosmos::CosmosMailbox::new(provider, conf.clone(), locator.clone())
+                h_cosmos::cw::CwMailbox::new(provider, conf.clone(), locator.clone())
                     .map(|m| Box::new(m) as Box<dyn Mailbox>)
                     .map_err(Into::into)
             }
@@ -377,9 +375,9 @@ impl ChainConf {
                     .map_err(Into::into)
             }
             ChainConnectionConf::CosmosNative(conf) => {
-                let signer = self.cosmos_native_signer().await.context(ctx)?;
+                let signer = self.cosmos_signer().await.context(ctx)?;
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, signer)?;
-                h_cosmos_native::CosmosNativeMailbox::new(provider, locator.clone())
+                h_cosmos::native::CosmosNativeMailbox::new(provider, locator.clone())
                     .map(|m| Box::new(m) as Box<dyn Mailbox>)
                     .map_err(Into::into)
             }
@@ -421,8 +419,8 @@ impl ChainConf {
             }
             ChainConnectionConf::Cosmos(conf) => {
                 let signer = self.cosmos_signer().await.context(ctx)?;
-                let provider = build_cosmos_provider(self, conf, metrics, &locator, signer)?;
-                let hook = h_cosmos::CosmosMerkleTreeHook::new(provider, locator.clone())?;
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, signer)?;
+                let hook = h_cosmos::cw::CwMerkleTreeHook::new(provider, locator.clone())?;
 
                 Ok(Box::new(hook) as Box<dyn MerkleTreeHook>)
             }
@@ -433,7 +431,7 @@ impl ChainConf {
             ChainConnectionConf::CosmosNative(conf) => {
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, None)?;
                 let hook =
-                    h_cosmos_native::CosmosNativeMerkleTreeHook::new(provider, locator.clone())?;
+                    h_cosmos::native::CosmosNativeMerkleTreeHook::new(provider, locator.clone())?;
 
                 Ok(Box::new(hook) as Box<dyn MerkleTreeHook>)
             }
@@ -485,26 +483,13 @@ impl ChainConf {
                 Ok(indexer as Box<dyn SequenceAwareIndexer<HyperlaneMessage>>)
             }
             ChainConnectionConf::Cosmos(conf) => {
-                let signer = self.cosmos_signer().await.context(ctx)?;
-                let reorg_period = self.reorg_period.as_blocks().context(ctx)?;
-
-                let provider =
-                    build_cosmos_provider(self, conf, metrics, &locator, signer.clone())?;
-                let wasm_provider = build_cosmos_wasm_provider(
-                    self,
-                    conf,
-                    &locator,
-                    metrics,
-                    reorg_period,
-                    dispatch_indexer::MESSAGE_DISPATCH_EVENT_TYPE.into(),
-                )?;
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, None)?;
 
                 let mailbox =
-                    h_cosmos::CosmosMailbox::new(provider, conf.clone(), locator.clone())?;
+                    h_cosmos::cw::CwMailbox::new(provider.clone(), conf.clone(), locator.clone())?;
                 let indexer = Box::new(
-                    h_cosmos::dispatch_indexer::CosmosMailboxDispatchIndexer::new(
-                        wasm_provider,
-                        mailbox,
+                    h_cosmos::cw::dispatch_indexer::CwMailboxDispatchIndexer::new(
+                        provider, mailbox, &locator,
                     )?,
                 );
                 Ok(indexer as Box<dyn SequenceAwareIndexer<HyperlaneMessage>>)
@@ -519,7 +504,7 @@ impl ChainConf {
             }
             ChainConnectionConf::CosmosNative(conf) => {
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, None)?;
-                let indexer = Box::new(h_cosmos_native::CosmosNativeDispatchIndexer::new(
+                let indexer = Box::new(h_cosmos::native::CosmosNativeDispatchIndexer::new(
                     provider, locator,
                 )?);
                 Ok(indexer as Box<dyn SequenceAwareIndexer<HyperlaneMessage>>)
@@ -572,18 +557,11 @@ impl ChainConf {
                 Ok(indexer as Box<dyn SequenceAwareIndexer<H256>>)
             }
             ChainConnectionConf::Cosmos(conf) => {
-                let reorg_period = self.reorg_period.as_blocks().context(ctx)?;
-                let wasm_provider = build_cosmos_wasm_provider(
-                    self,
-                    conf,
-                    &locator,
-                    metrics,
-                    reorg_period,
-                    delivery_indexer::MESSAGE_DELIVERY_EVENT_TYPE.into(),
-                )?;
-
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, None)?;
                 let indexer = Box::new(
-                    h_cosmos::delivery_indexer::CosmosMailboxDeliveryIndexer::new(wasm_provider)?,
+                    h_cosmos::cw::delivery_indexer::CwMailboxDeliveryIndexer::new(
+                        provider, &locator,
+                    ),
                 );
                 Ok(indexer as Box<dyn SequenceAwareIndexer<H256>>)
             }
@@ -597,7 +575,7 @@ impl ChainConf {
             }
             ChainConnectionConf::CosmosNative(conf) => {
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, None)?;
-                let indexer = Box::new(h_cosmos_native::CosmosNativeDeliveryIndexer::new(
+                let indexer = Box::new(h_cosmos::native::CosmosNativeDeliveryIndexer::new(
                     provider, locator,
                 )?);
                 Ok(indexer as Box<dyn SequenceAwareIndexer<H256>>)
@@ -643,9 +621,9 @@ impl ChainConf {
             }
             ChainConnectionConf::Cosmos(conf) => {
                 let signer = self.cosmos_signer().await.context(ctx)?;
-                let provider = build_cosmos_provider(self, conf, metrics, &locator, signer)?;
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, signer)?;
 
-                let paymaster = Box::new(h_cosmos::CosmosInterchainGasPaymaster::new(
+                let paymaster = Box::new(h_cosmos::cw::CwInterchainGasPaymaster::new(
                     provider,
                     locator.clone(),
                 )?);
@@ -659,7 +637,7 @@ impl ChainConf {
             }
             ChainConnectionConf::CosmosNative(conf) => {
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, None)?;
-                let indexer = Box::new(h_cosmos_native::CosmosNativeInterchainGas::new(
+                let indexer = Box::new(h_cosmos::native::CosmosNativeInterchainGas::new(
                     provider, conf, locator,
                 )?);
                 Ok(indexer as Box<dyn InterchainGasPaymaster>)
@@ -701,7 +679,8 @@ impl ChainConf {
             }
             ChainConnectionConf::Fuel(_) => todo!(),
             ChainConnectionConf::Sealevel(conf) => {
-                let provider = Arc::new(build_sealevel_provider(self, &locator, &[], conf, metrics));
+                let provider =
+                    Arc::new(build_sealevel_provider(self, &locator, &[], conf, metrics));
                 let indexer = Box::new(
                     h_sealevel::SealevelInterchainGasPaymasterIndexer::new(
                         provider,
@@ -713,19 +692,11 @@ impl ChainConf {
                 Ok(indexer as Box<dyn SequenceAwareIndexer<InterchainGasPayment>>)
             }
             ChainConnectionConf::Cosmos(conf) => {
-                let reorg_period = self.reorg_period.as_blocks().context(ctx)?;
-                let wasm_provider = build_cosmos_wasm_provider(
-                    self,
-                    conf,
-                    &locator,
-                    metrics,
-                    reorg_period,
-                    h_cosmos::CosmosInterchainGasPaymasterIndexer::INTERCHAIN_GAS_PAYMENT_EVENT_TYPE.into(),
-                )?;
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, None)?;
 
-                let indexer = Box::new(h_cosmos::CosmosInterchainGasPaymasterIndexer::new(
-                    wasm_provider,
-                )?);
+                let indexer = Box::new(h_cosmos::cw::CwInterchainGasPaymasterIndexer::new(
+                    provider, &locator,
+                ));
                 Ok(indexer as Box<dyn SequenceAwareIndexer<InterchainGasPayment>>)
             }
             ChainConnectionConf::Starknet(_) => {
@@ -734,10 +705,8 @@ impl ChainConf {
             }
             ChainConnectionConf::CosmosNative(conf) => {
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, None)?;
-                let indexer = Box::new(h_cosmos_native::CosmosNativeInterchainGas::new(
-                    provider,
-                    conf,
-                    locator,
+                let indexer = Box::new(h_cosmos::native::CosmosNativeInterchainGas::new(
+                    provider, conf, locator,
                 )?);
                 Ok(indexer as Box<dyn SequenceAwareIndexer<InterchainGasPayment>>)
             }
@@ -794,22 +763,10 @@ impl ChainConf {
             }
             ChainConnectionConf::Cosmos(conf) => {
                 let signer = self.cosmos_signer().await.context(ctx)?;
-                let reorg_period = self.reorg_period.as_blocks().context(ctx)?;
 
-                let provider = build_cosmos_provider(self, conf, metrics, &locator, signer)?;
-                let wasm_provider = build_cosmos_wasm_provider(
-                    self,
-                    conf,
-                    &locator,
-                    metrics,
-                    reorg_period,
-                    h_cosmos::CosmosMerkleTreeHookIndexer::MERKLE_TREE_INSERTION_EVENT_TYPE.into(),
-                )?;
-
-                let indexer = Box::new(h_cosmos::CosmosMerkleTreeHookIndexer::new(
-                    provider,
-                    wasm_provider,
-                    locator,
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, signer)?;
+                let indexer = Box::new(h_cosmos::cw::CwMerkleTreeHookIndexer::new(
+                    provider, locator,
                 )?);
                 Ok(indexer as Box<dyn SequenceAwareIndexer<MerkleTreeInsertion>>)
             }
@@ -823,7 +780,7 @@ impl ChainConf {
             }
             ChainConnectionConf::CosmosNative(conf) => {
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, None)?;
-                let indexer = Box::new(h_cosmos_native::CosmosNativeMerkleTreeHook::new(
+                let indexer = Box::new(h_cosmos::native::CosmosNativeMerkleTreeHook::new(
                     provider, locator,
                 )?);
                 Ok(indexer as Box<dyn SequenceAwareIndexer<MerkleTreeInsertion>>)
@@ -862,9 +819,9 @@ impl ChainConf {
             }
             ChainConnectionConf::Cosmos(conf) => {
                 let signer = self.cosmos_signer().await.context(ctx)?;
-                let provider = build_cosmos_provider(self, conf, metrics, &locator, signer)?;
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, signer)?;
 
-                let va = Box::new(h_cosmos::CosmosValidatorAnnounce::new(
+                let va = Box::new(h_cosmos::cw::CwValidatorAnnounce::new(
                     provider,
                     locator.clone(),
                 )?);
@@ -880,9 +837,9 @@ impl ChainConf {
                 Ok(va as Box<dyn ValidatorAnnounce>)
             }
             ChainConnectionConf::CosmosNative(conf) => {
-                let signer = self.cosmos_native_signer().await.context(ctx)?;
+                let signer = self.cosmos_signer().await.context(ctx)?;
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, signer)?;
-                let va = Box::new(h_cosmos_native::CosmosNativeValidatorAnnounce::new(
+                let va = Box::new(h_cosmos::native::CosmosNativeValidatorAnnounce::new(
                     provider,
                     locator.clone(),
                 )?);
@@ -934,9 +891,9 @@ impl ChainConf {
             }
             ChainConnectionConf::Cosmos(conf) => {
                 let signer = self.cosmos_signer().await.context(ctx)?;
-                let provider = build_cosmos_provider(self, conf, metrics, &locator, signer)?;
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, signer)?;
 
-                let ism = Box::new(h_cosmos::CosmosInterchainSecurityModule::new(
+                let ism = Box::new(h_cosmos::cw::CwInterchainSecurityModule::new(
                     provider, locator,
                 )?);
                 Ok(ism as Box<dyn InterchainSecurityModule>)
@@ -949,7 +906,7 @@ impl ChainConf {
             }
             ChainConnectionConf::CosmosNative(conf) => {
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, None)?;
-                let ism = Box::new(h_cosmos_native::CosmosNativeIsm::new(provider, locator)?);
+                let ism = Box::new(h_cosmos::native::CosmosNativeIsm::new(provider, locator)?);
                 Ok(ism as Box<dyn InterchainSecurityModule>)
             }
             ChainConnectionConf::Radix(conf) => {
@@ -990,9 +947,9 @@ impl ChainConf {
             }
             ChainConnectionConf::Cosmos(conf) => {
                 let signer = self.cosmos_signer().await.context(ctx)?;
-                let provider = build_cosmos_provider(self, conf, metrics, &locator, signer)?;
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, signer)?;
 
-                let ism = Box::new(h_cosmos::CosmosMultisigIsm::new(provider, locator.clone())?);
+                let ism = Box::new(h_cosmos::cw::CwMultisigIsm::new(provider, locator.clone())?);
                 Ok(ism as Box<dyn MultisigIsm>)
             }
             ChainConnectionConf::Starknet(conf) => {
@@ -1001,8 +958,7 @@ impl ChainConf {
             }
             ChainConnectionConf::CosmosNative(conf) => {
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, None)?;
-                let ism: Box<hyperlane_cosmos_native::CosmosNativeIsm> =
-                    Box::new(h_cosmos_native::CosmosNativeIsm::new(provider, locator)?);
+                let ism = Box::new(h_cosmos::native::CosmosNativeIsm::new(provider, locator)?);
                 Ok(ism as Box<dyn MultisigIsm>)
             }
             ChainConnectionConf::Radix(conf) => {
@@ -1037,9 +993,9 @@ impl ChainConf {
             }
             ChainConnectionConf::Cosmos(conf) => {
                 let signer = self.cosmos_signer().await.context(ctx)?;
-                let provider = build_cosmos_provider(self, conf, metrics, &locator, signer)?;
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, signer)?;
 
-                let ism = Box::new(h_cosmos::CosmosRoutingIsm::new(provider, locator.clone())?);
+                let ism = Box::new(h_cosmos::cw::CwRoutingIsm::new(provider, locator.clone())?);
                 Ok(ism as Box<dyn RoutingIsm>)
             }
             ChainConnectionConf::Starknet(conf) => {
@@ -1048,8 +1004,7 @@ impl ChainConf {
             }
             ChainConnectionConf::CosmosNative(conf) => {
                 let provider = build_cosmos_native_provider(self, conf, metrics, &locator, None)?;
-                let ism: Box<hyperlane_cosmos_native::CosmosNativeIsm> =
-                    Box::new(h_cosmos_native::CosmosNativeIsm::new(provider, locator)?);
+                let ism = Box::new(h_cosmos::native::CosmosNativeIsm::new(provider, locator)?);
                 Ok(ism as Box<dyn RoutingIsm>)
             }
             ChainConnectionConf::Radix(conf) => {
@@ -1084,8 +1039,8 @@ impl ChainConf {
             }
             ChainConnectionConf::Cosmos(conf) => {
                 let signer = self.cosmos_signer().await.context(ctx)?;
-                let provider = build_cosmos_provider(self, conf, metrics, &locator, signer)?;
-                let ism = Box::new(h_cosmos::CosmosAggregationIsm::new(
+                let provider = build_cosmos_wasm_provider(self, conf, metrics, &locator, signer)?;
+                let ism = Box::new(h_cosmos::cw::CwAggregationIsm::new(
                     provider,
                     locator.clone(),
                 )?);
@@ -1164,12 +1119,11 @@ impl ChainConf {
                 ChainConnectionConf::Sealevel(_) => {
                     Box::new(conf.build::<h_sealevel::Keypair>().await?)
                 }
-                ChainConnectionConf::Cosmos(_) => Box::new(conf.build::<h_cosmos::Signer>().await?),
+                ChainConnectionConf::Cosmos(_) | ChainConnectionConf::CosmosNative(_) => {
+                    Box::new(conf.build::<h_cosmos::Signer>().await?)
+                }
                 ChainConnectionConf::Starknet(_) => {
                     Box::new(conf.build::<h_starknet::Signer>().await?)
-                }
-                ChainConnectionConf::CosmosNative(_) => {
-                    Box::new(conf.build::<h_cosmos_native::Signer>().await?)
                 }
                 ChainConnectionConf::Radix(_) => {
                     Box::new(conf.build::<h_radix::RadixSigner>().await?)
@@ -1200,10 +1154,6 @@ impl ChainConf {
     }
 
     async fn starknet_signer(&self) -> Result<Option<h_starknet::Signer>> {
-        self.signer().await
-    }
-
-    async fn cosmos_native_signer(&self) -> Result<Option<h_cosmos_native::Signer>> {
         self.signer().await
     }
 
@@ -1338,57 +1288,34 @@ fn build_sealevel_tx_submitter(
     )
 }
 
-fn build_cosmos_provider(
-    chain_conf: &ChainConf,
-    connection_conf: &h_cosmos::ConnectionConf,
-    metrics: &CoreMetrics,
-    locator: &ContractLocator,
-    signer: Option<Signer>,
-) -> ChainResult<CosmosProvider> {
-    let middleware_metrics = chain_conf.metrics_conf();
-    let client_metrics = metrics.client_metrics();
-
-    CosmosProvider::new(
-        locator.domain.clone(),
-        connection_conf.clone(),
-        locator,
-        signer,
-        client_metrics,
-        middleware_metrics.chain.clone(),
-    )
-}
-
 fn build_cosmos_wasm_provider(
     chain_conf: &ChainConf,
     connection_conf: &h_cosmos::ConnectionConf,
-    locator: &ContractLocator,
     metrics: &CoreMetrics,
-    reorg_period: u32,
-    event_type: String,
-) -> ChainResult<CosmosWasmRpcProvider> {
+    locator: &ContractLocator,
+    signer: Option<hyperlane_cosmos::Signer>,
+) -> ChainResult<CosmosProvider<CwQueryClient>> {
     let middleware_metrics = chain_conf.metrics_conf();
-    let client_metrics = metrics.client_metrics();
-
-    CosmosWasmRpcProvider::new(
+    let metrics = metrics.client_metrics();
+    CosmosProvider::new(
         connection_conf,
         locator,
-        event_type,
-        reorg_period,
-        client_metrics,
+        signer,
+        metrics,
         middleware_metrics.chain.clone(),
     )
 }
 
 fn build_cosmos_native_provider(
     chain_conf: &ChainConf,
-    connection_conf: &h_cosmos_native::ConnectionConf,
+    connection_conf: &h_cosmos::ConnectionConf,
     metrics: &CoreMetrics,
     locator: &ContractLocator,
-    signer: Option<hyperlane_cosmos_native::Signer>,
-) -> ChainResult<CosmosNativeProvider> {
+    signer: Option<hyperlane_cosmos::Signer>,
+) -> ChainResult<CosmosProvider<ModuleQueryClient>> {
     let middleware_metrics = chain_conf.metrics_conf();
     let metrics = metrics.client_metrics();
-    CosmosNativeProvider::new(
+    CosmosProvider::new(
         connection_conf,
         locator,
         signer,