+++ typescript/infra/config/environments/mainnet3/agent.ts
@@ -69,7 +69,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     abstract: true,
     // acala: true,
     ancient8: true,
-    alephzeroevmmainnet: true,
+    alephzeroevmmainnet: false,
     apechain: true,
     appchain: true,
     arbitrum: true,
@@ -99,6 +99,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     degenchain: true,
     dogechain: true,
     eclipsemainnet: true,
+    electroneum: true,
     endurance: true,
     ethereum: true,
     everclear: true,
@@ -111,7 +112,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     fraxtal: true,
     fusemainnet: true,
     galactica: true,
-    game7: true,
+    game7: false,
     gnosis: true,
     gravity: true,
     harmony: true,
@@ -129,7 +130,6 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     linea: true,
     lisk: true,
     lukso: true,
-    lumia: true,
     lumiaprism: true,
     mantapacific: true,
     mantle: true,
@@ -156,6 +156,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     osmosis: true,
     paradex: true,
     peaq: true,
+    plasma: true,
     plume: true,
     polygon: true,
     polygonzkevm: true,
@@ -208,7 +209,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     abstract: true,
     // acala: true,
     ancient8: true,
-    alephzeroevmmainnet: true,
+    alephzeroevmmainnet: false,
     apechain: true,
     appchain: true,
     arcadia: true,
@@ -238,6 +239,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     degenchain: true,
     dogechain: true,
     eclipsemainnet: true,
+    electroneum: true,
     endurance: true,
     ethereum: true,
     everclear: true,
@@ -250,7 +252,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     fraxtal: true,
     fusemainnet: true,
     galactica: true,
-    game7: true,
+    game7: false,
     gnosis: true,
     gravity: true,
     harmony: true,
@@ -268,7 +270,6 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     linea: true,
     lisk: true,
     lukso: true,
-    lumia: true,
     lumiaprism: true,
     mantapacific: true,
     mantle: true,
@@ -295,6 +296,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     osmosis: true,
     paradex: true,
     peaq: true,
+    plasma: true,
     plume: true,
     polygon: true,
     polygonzkevm: true,
@@ -339,7 +341,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     xrplevm: true,
     zeronetwork: true,
     zetachain: true,
-    zircuit: false,
+    zircuit: true,
     zksync: true,
     zoramainnet: true,
   },
@@ -347,7 +349,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     abstract: true,
     // acala: true,
     ancient8: true,
-    alephzeroevmmainnet: true,
+    alephzeroevmmainnet: false,
     apechain: true,
     appchain: true,
     arbitrum: true,
@@ -377,6 +379,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     degenchain: true,
     dogechain: true,
     eclipsemainnet: true,
+    electroneum: true,
     endurance: true,
     ethereum: true,
     everclear: true,
@@ -389,7 +392,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     fraxtal: true,
     fusemainnet: true,
     galactica: true,
-    game7: true,
+    game7: false,
     gnosis: true,
     gravity: true,
     harmony: true,
@@ -407,7 +410,6 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     linea: true,
     lisk: true,
     lukso: true,
-    lumia: true,
     lumiaprism: true,
     mantapacific: true,
     mantle: true,
@@ -434,6 +436,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     osmosis: true,
     paradex: true,
     peaq: true,
+    plasma: true,
     plume: true,
     polygon: true,
     polygonzkevm: true,
@@ -441,7 +444,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     prom: true,
     proofofplay: true,
     pulsechain: true,
-    radix: false,
+    radix: true,
     rarichain: true,
     reactive: true,
     redstone: true,
@@ -479,7 +482,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     xrplevm: true,
     zeronetwork: true,
     zetachain: true,
-    zircuit: false,
+    zircuit: true,
     zksync: true,
     zoramainnet: true,
   },
@@ -651,13 +654,6 @@ const stagingStHyperMatchingList = chainMapMatchingList({
   ethereum: '0x0C919509663cb273E156B706f065b9F7e6331891',
 });
 
-const vanguardMatchingList = [
-  ...hyperMatchingList,
-  ...stHyperMatchingList,
-  ...stagingHyperMatchingList,
-  ...stagingStHyperMatchingList,
-];
-
 // Gets metric app contexts, including:
 // - helloworld
 // - all warp routes defined in WarpRouteIds, using addresses from the registry
@@ -804,6 +800,11 @@ const blacklist: MatchingList = [
       '0x867c1fd9341DEC12e4B779C35D7b7C475316b334',
     ],
   },
+  // StarkNet<>StarkNet messages
+  {
+    originDomain: getDomainId('starknet'),
+    destinationDomain: getDomainId('starknet'),
+  },
 ];
 
 const ismCacheConfigs: Array<IsmCacheConfig> = [
@@ -834,7 +835,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '7141ffc-20250828-121236',
+      tag: 'a62b4a4-20250918-104420',
     },
     blacklist,
     gasPaymentEnforcement: gasPaymentEnforcement,
@@ -854,7 +855,7 @@ const hyperlane: RootAgentConfig = {
   validators: {
     docker: {
       repo,
-      tag: '7141ffc-20250828-121236',
+      tag: '093e0be-20250904-212216',
     },
     rpcConsensusType: RpcConsensusType.Quorum,
     chains: validatorChainConfig(Contexts.Hyperlane),
@@ -865,7 +866,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '7141ffc-20250828-121236',
+      tag: 'e8b16e1-20250914-163558',
     },
     resources: scraperResources,
   },
@@ -880,7 +881,7 @@ const releaseCandidate: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '7141ffc-20250828-121236',
+      tag: 'a62b4a4-20250918-104420',
     },
     blacklist,
     // We're temporarily (ab)using the RC relayer as a way to increase
@@ -903,7 +904,7 @@ const releaseCandidate: RootAgentConfig = {
   validators: {
     docker: {
       repo,
-      tag: '7141ffc-20250828-121236',
+      tag: '093e0be-20250904-212216',
     },
     rpcConsensusType: RpcConsensusType.Quorum,
     chains: validatorChainConfig(Contexts.ReleaseCandidate),
@@ -943,79 +944,8 @@ const neutron: RootAgentConfig = {
   },
 };
 
-const getVanguardRootAgentConfig = (index: number): RootAgentConfig => ({
-  ...contextBase,
-  context: mustBeValidContext(`vanguard${index}`),
-  contextChainNames: {
-    validator: [],
-    relayer: ['bsc', 'arbitrum', 'optimism', 'ethereum', 'base'],
-    scraper: [],
-  },
-  rolesWithKeys: [Role.Relayer],
-  relayer: {
-    rpcConsensusType: RpcConsensusType.Fallback,
-    docker: {
-      repo,
-      // includes gasPriceCap overrides + per-chain maxSubmitQueueLength
-      tag: '420c950-20250612-172436',
-    },
-    whitelist: vanguardMatchingList,
-    // Not specifying a blacklist for optimization purposes -- all the message IDs
-    // in there are not vanguard-specific.
-    gasPaymentEnforcement: [
-      {
-        type: GasPaymentEnforcementPolicyType.None,
-        matchingList: vanguardMatchingList,
-      },
-    ],
-    metricAppContextsGetter,
-    ismCacheConfigs,
-    cache: {
-      enabled: true,
-      // Cache for 10 minutes
-      defaultExpirationSeconds: 10 * 60,
-    },
-    resources: {
-      requests: {
-        // Big enough to claim a c3-standard-44 each
-        cpu: '35000m',
-        memory: '100Gi',
-      },
-    },
-    dbBootstrap: true,
-    mixing: {
-      enabled: true,
-      // Arbitrary salt to ensure different agents have different sorting behavior for pending messages
-      salt: 69690 + index,
-    },
-    batch: {
-      defaultBatchSize: 32,
-      batchSizeOverrides: {
-        // Slightly lower to ideally fit within 5M
-        ethereum: 26,
-      },
-      bypassBatchSimulation: true,
-      maxSubmitQueueLength: {
-        arbitrum: 350,
-        base: 350,
-        bsc: 350,
-        optimism: 350,
-        ethereum: 75,
-      },
-    },
-    txIdIndexingEnabled: false,
-    igpIndexingEnabled: false,
-  },
-});
-
 export const agents = {
   [Contexts.Hyperlane]: hyperlane,
   [Contexts.ReleaseCandidate]: releaseCandidate,
   [Contexts.Neutron]: neutron,
-  [Contexts.Vanguard0]: getVanguardRootAgentConfig(0),
-  [Contexts.Vanguard1]: getVanguardRootAgentConfig(1),
-  [Contexts.Vanguard2]: getVanguardRootAgentConfig(2),
-  [Contexts.Vanguard3]: getVanguardRootAgentConfig(3),
-  [Contexts.Vanguard4]: getVanguardRootAgentConfig(4),
-  [Contexts.Vanguard5]: getVanguardRootAgentConfig(5),
 };