+++ typescript/infra/config/environments/mainnet3/customBlacklist.ts
@@ -175,4 +175,8 @@ export const blacklistedMessageIds = [
   // Paradex USDC:
   // 8/27/2025
   '0x151135eda5f04110084c6673d23f041d9141a50bdb02de028e32fe7f4e14e7c2',
+
+  // 09/09/2025
+  // Our own message on SVM that references our own relayer address
+  '0x337d60f17842a6b152f52f41ead0fc6c5f944543a7f15b1c7d31e35dcb943309',
 ];