+++ typescript/infra/config/environments/testnet4/funding.ts
@@ -10,7 +10,7 @@ export const keyFunderConfig: KeyFunderConfig<
 > = {
   docker: {
     repo: 'gcr.io/abacus-labs-dev/hyperlane-monorepo',
-    tag: '7dca2ae-20250825-162920',
+    tag: 'e5e9700-20250909-164327',
   },
   // We're currently using the same deployer key as testnet2.
   // To minimize nonce clobbering we offset the key funder cron
@@ -27,34 +27,24 @@ export const keyFunderConfig: KeyFunderConfig<
   chainsToSkip: ['hyperliquidevmtestnet', 'infinityvmmonza'],
   // desired balance config
   desiredBalancePerChain: {
-    abstracttestnet: '0.1',
-    alephzeroevmtestnet: '2',
-    alfajores: '5',
     arbitrumsepolia: '0.1',
     arcadiatestnet2: '0.1',
     auroratestnet: '0.05',
     basecamptestnet: '0.05',
     basesepolia: '0.1',
-    bepolia: '0.05',
     bsctestnet: '5',
     carrchaintestnet: '100',
     celestiatestnet: '0',
     celosepolia: '0.5',
-    chronicleyellowstone: '0.001',
     citreatestnet: '0.001',
-    connextsepolia: '1',
     cotitestnet: '1',
-    ecotestnet: '0.02',
     // no funding for solana
     eclipsetestnet: '0',
-    flametestnet: '0.1',
-    formtestnet: '0.1',
     fuji: '5',
     holesky: '5',
     hyperliquidevmtestnet: '0.1',
     incentivtestnet: '1',
     infinityvmmonza: '0',
-    inksepolia: '0.1',
     kyvetestnet: '0',
     megaethtestnet: '0.01',
     milkywaytestnet: '0',
@@ -62,10 +52,8 @@ export const keyFunderConfig: KeyFunderConfig<
     monadtestnet: '0.1',
     neuratestnet: '0.1',
     nobletestnet: '0',
-    odysseytestnet: '0.1',
     optimismsepolia: '0.1',
     paradexsepolia: '0',
-    plumetestnet2: '0.1',
     polygonamoy: '0.2',
     radixtestnet: '0',
     scrollsepolia: '1',
@@ -74,17 +62,11 @@ export const keyFunderConfig: KeyFunderConfig<
     // no funding for SVM chains
     solanatestnet: '0',
     somniatestnet: '10',
-    soneiumtestnet: '0.1',
-    sonicblaze: '0.1',
     // no funding for SVM chains
     sonicsvmtestnet: '0',
     subtensortestnet: '0.1',
-    superpositiontestnet: '1',
-    unichaintestnet: '0.1',
-    weavevmtestnet: '0.1',
   },
   desiredKathyBalancePerChain: {
-    alfajores: '1',
     arbitrumsepolia: '0',
     basesepolia: '0',
     bsctestnet: '1',
@@ -105,7 +87,6 @@ export const keyFunderConfig: KeyFunderConfig<
     superpositiontestnet: '0',
   },
   igpClaimThresholdPerChain: {
-    alfajores: '1',
     arbitrumsepolia: '0.05',
     basesepolia: '0.05',
     bsctestnet: '1',